ref: make API thinner by moving simple calls into EngineGetParm(RenderGetParm with ref extensions), fix some UBs

+++ common/render_api.h
@@ -235,7 +235,8 @@ typedef struct render_api_s
 	float		(*pfnTime)( void );				// Sys_DoubleTime
 	void		(*Cvar_Set)( const char *name, const char *value );
 	void		(*S_FadeMusicVolume)( float fadePercent );	// fade background track (0-100 percents)
-	void		(*SetRandomSeed)( long lSeed );		// set custom seed for RANDOM_FLOAT\RANDOM_LONG for predictable random
+	// a1ba: changed long to int
+	void		(*SetRandomSeed)( int lSeed );		// set custom seed for RANDOM_FLOAT\RANDOM_LONG for predictable random
 	// ONLY ADD NEW FUNCTIONS TO THE END OF THIS STRUCT.  INTERFACE VERSION IS FROZEN AT 37
 } render_api_t;
